+++ derive/src/utils/impl_block/from_impl_item_method.rs
@@ -1,11 +1,11 @@
 use darling::util::Ignored;
 
-pub trait FromImplItemMethod: Sized {
-    fn from_impl_item_method(impl_item_method: &mut syn::ImplItemMethod) -> darling::Result<Self>;
+pub trait FromImplItemFn: Sized {
+    fn from_impl_item_method(impl_item_method: &mut syn::ImplItemFn) -> darling::Result<Self>;
 }
 
-impl FromImplItemMethod for Ignored {
-    fn from_impl_item_method(_method: &mut syn::ImplItemMethod) -> darling::Result<Self> {
+impl FromImplItemFn for Ignored {
+    fn from_impl_item_method(_method: &mut syn::ImplItemFn) -> darling::Result<Self> {
         Ok(Ignored)
     }
 }