+++ derive/src/utils/meta_match/match_nested_meta.rs
@@ -1,3 +1,5 @@
+use darling::ast::NestedMeta;
+
 pub trait MatchNestedMeta: Sized {
-    fn match_nested_meta(meta: &syn::NestedMeta) -> Option<darling::Result<Self>>;
+    fn match_nested_meta(meta: &NestedMeta) -> Option<darling::Result<Self>>;
 }