+++ derive/src/utils/register_attr.rs
@@ -1,9 +1,9 @@
+use darling::ast::NestedMeta;
 use darling::FromMeta;
 use proc_macro2::Span;
 use quote::quote;
 use quote::ToTokens;
 use syn::spanned::Spanned;
-use syn::NestedMeta;
 
 use crate::utils::meta_match::MatchMetaPath;
 use crate::utils::meta_match::MatchNestedMetaList;